IFC-FSAR now functional
Re-coded IFC stock-recruit analysis in Stan. All residuals and parameter outputs have been checked in plots. All code should run as long as the user open the R project and then runs the .Rmd file. Note that the .rds files (the mode fits) are not included in the repository as they are too large to store in Github, but the code in the .Rmd doc will create new .rds files.